refactor: split additionalGames.ts into 6 domain files + barrel (#477)
Moves medicine, natureSounds, seasonsHolidays, feelings, shoppingMoney,
roadSafety into additionalGamesData/ subfiles. additionalGames.ts becomes
a 6-line barrel — zero breaking changes for consumers.
